+++ app/Models/NetworkImpression.php
@@ -44,6 +44,7 @@ use function hex2bin;
  * @property float|null human_score
  * @property float|null page_rank
  * @property string|null user_data
+ * @property string|null country
  * @property Collection networkCases
  * @mixin Builder
  */
@@ -107,6 +108,7 @@ class NetworkImpression extends Model
         $this->human_score = $userContext->humanScore();
         $this->page_rank = $userContext->pageRank();
         $this->user_data = $userContext->keywords();
+        $this->country = $userContext->country();
     }
 
     public function networkCases(): HasMany