+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Send an email notification once the banner was classified
 - Reassessment of distinctive sites
 - Sites' categories
+- Server's statistics backup
 ### Changed
 - Do not allow site's domain starts with a dot
 - Taxonomy processing, allow multiple levels